Fixed Issue with "resource grid".
I was having an issue with the code for the "resources grid" section on my home page, mainly with using CSS grid to formate the section. Found that the issue was originating from my use of classes and have fixed that. Next I need to finish the formatting and have to add icons.