Finished Base code of review section, started created footer
Finished the base code of the review section of my home page, need to replace the filler text with actual text. Started creating the footer, managed to get the layout the way I wanted it and coloured everything accordinly. Need to finish by linking the empty HTML pages and the anchor tag. Also need to add the icons to the titles of the lists.