Created empty HTML pages, Finished footer.
I have finished creating the footer. Decided that adding icons would over crowd the footer and would make it look tacky. Add 2 media quires conditions for the footer to make it more acsessible on all devices. Lastly, fixed the issue I was having with the nav bar and the issue that I was having with the responsiveness of Home-page main section.