Hero Images + Nav bar + Footers
I have added hero images to all the pages (apart from the about page and the contact page). Thus I have started the HTML for the majority of the pages including nav bars and footers. Having issues with the margin on the subitems as it makes it difficult for the users to presses the link they want. However, I will be attempting to fix this soon.