Nav bar bug and Maths resource page
I managed to fix the navigation bar issue by increasing the margin and padding around the submenu items, making it easier for the user to use. I also created a hero image on all pages and pasted in the basic code for the resource page layout. I have almost completed the maths resource page.