Started making the chemistry page.
Started making the chemistry page, by linking videos, past papers and asessment schecules. Also made the resource document for the first standard on the chemistry page and started on the second standard's resource's doc and both have been embedded into the website. There is still a bug I need to fix revolving the embedded google document, this is discussed in more detail on the issue section of the projects management page.