Media quires for resource pages
I have added the CSS code for the resource pages' media quires, meaning that the user will be able to easisly use the website on a mobile device. Now, when the screen size is 800px or less, the resource section will convert to a layout where there is only one item on each row and they are stacked vertically.