Change of resource page layout + edits to embedded docs.
I have decided that for the user and the accessabilty of the website and the usabitly and asethic of the website, that I should change the layout of the resource section. I have now made it so that the docs only have a margin of 1 (on the actuclly google doc) and have set it so that the doc section take up the whole bottom  row and the past paper and video section share the row above.