edits to the embedded doc.
I have managed to set the google doc to be centered in the middle of the grid item, however this does come with limitations. If i try to set the width to of the embedded i frame to bigger the 50vw, the doc will resume its postion on the left hand side.  I did also change  the CSS code for the iframe, and also changed the docs iframe class.