New Resource page layout (maths, physics, chemistry)
I have edited the way the resource sections in the resource pages are laid out. Now instead of having the all the sections (past papers, video resources, written resources) displayed at all time, I have made it so that it only opens when the user presses on their title. This way the user is not overwhelmed by the amount of content open on the page, and they can now easily navigate their way to their standard of choice with out having to scroll through the content of all the other standards on the page. This change makes the mobile version of the website much more efficent to use.